% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{addq}
\alias{addq}
\title{Add question-answer}
\usage{
addq(tags = c(""))
}
\arguments{
\item{tags}{an optional string containing tags to classify question into a category(s)}
}
\value{
three objects: q_tbl, sol_tbl and data_tbl updated with the new question-answer.
}
\description{
Add question in one line (without carriage return). Add answers in the following line(s). Carriage return in a blank line will save the question and answer to the data objects.
}
\examples{
\dontrun{addq()}

}
