% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{bye}
\alias{bye}
\title{Closes the quiz session}
\usage{
bye()
}
\value{
None
}
\description{
This is important as it updates the files on disk with any new questions added in current session. After updating the qa file, it clears out the R environment by removing the objects and functions related to this package.
}
\examples{
\dontrun{bye()}

}
