% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{changeq}
\alias{changeq}
\title{Change question-answer}
\usage{
changeq()
}
\value{
two objects: q_tbl and sol_tbl updated with the new question-answer.
}
\description{
Change/replace the current question. First line will be the questio. Add answers in the following line(s). Carriage return in a blank line will save the question and answer to the data objects.
}
\examples{
\dontrun{changeq()}

}
