% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{delete_last}
\alias{delete_last}
\title{Delete last question}
\usage{
delete_last()
}
\value{
updated rank table, question and answer tables
}
\description{
deletes the last added question to the database.
}
\examples{
\dontrun{delete_last()}

}
