% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{hit}
\alias{hit}
\title{Correct answer response}
\usage{
hit()
}
\value{
does not return anything, just updates the datalog
}
\description{
This will log data about correct response
}
\examples{
\dontrun{done()}

}
