% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{mvPastdueToToday}
\alias{mvPastdueToToday}
\title{move past due questions to today}
\usage{
mvPastdueToToday(DT)
}
\description{
importFrom lubridate now today hour minute second
}
\details{
this function update the due date of past due questions to today
this is needed for new questions that are many days past due
so that if they are practiced today, they are updated to learning tomorrow
and not some day in the past
}
