% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{show_status}
\alias{show_status}
\title{Test session status}
\usage{
show_status()
}
\value{
number of questions remaining
}
\description{
Reports number of questions remaining
}
\examples{
\dontrun{show_status()}

}
