% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{t_learning}
\alias{t_learning}
\title{Number of days elapsed}
\usage{
t_learning(tlast, qid)
}
\description{
Finds number of hours elapsed between the last two askings of qid
Finds new time(hours) to update for learning questions given tlast and qid
This will be called in by updatetime()
}
