% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{week_ahead}
\alias{week_ahead}
\title{Due next 7 days}
\usage{
week_ahead()
}
\value{
number of questions due in the next 7 days
}
\description{
Reports number of questions due in the next 7 days
}
\examples{
\dontrun{week_ahead()}

}
