% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{ask}
\alias{ask}
\title{Show a question from the Q&A repository}
\usage{
ask()
}
\value{
randomly selected question
}
\description{
A randomly selected question will be shown. For R questions, you can simply enter your answer (code) at the R console directly. Answers are not evaluated by the package. This is up to the user to decide whether they answered the question correctly or not.
}
\examples{
\dontrun{ask()}

}
