% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{tell}
\alias{tell}
\title{Show answer}
\usage{
tell()
}
\value{
answer to the last question presented. This helps the user to evaluate their answer.
}
\description{
This will show the answer to the last question presented
}
\examples{
\dontrun{tell()}

}
